Fixed a couple of issues...
Putting the PhysicalDamange and MagicalDamage string table decfinitions in utilities had them being included everywhere

+++ C-Practice/ElementalAffinity.cpp
@@ -2,7 +2,7 @@
 
 void ElementalAffinity::SetAffinityStrName(DamageTypes element)
 {
-	affinityName = DamageTypeStrNames[element];
+	affinityName = DamageTypeStrNames[static_cast<int>(element)];
 }
 
 void ElementalAffinity::SetElementalAffinity(int new_value)
